src/lib/core/datetime/native-date-adapter.ts
import { Inject, Injectable, Optional } from '@angular/core';

import { Platform } from '@ptsecurity/cdk/platform';

import { DateAdapter, MC_DATE_LOCALE } from './date-adapter';


const DAY_IN_MONTH = 31;
const DAY_IN_WEEK = 7;
const MONTH_IN_YEAR = 12;
const MAX_YEAR = 9999;

// Only need for calc months
const ANY_YEAR = 2018;

/**
 * Matches strings that have the form of a valid RFC 3339 string
 * (https://tools.ietf.org/html/rfc3339). Note that the string may not actually be a valid date
 * because the regex will match strings an with out of bounds month, date, etc.
 */
const ISO_8601_REGEX =
    /^\d{4}-\d{2}-\d{2}(?:T\d{2}:\d{2}:\d{2}(?:\.\d+)?(?:Z|(?:(?:\+|-)\d{2}:\d{2}))?)?$/;

/** Creates an array and fills it with values. */
function range<T>(length: number, valueFunction: (index: number) => T): T[] {
    const valuesArray = Array(length);

    for (let i = 0; i < length; i++) {
        valuesArray[i] = valueFunction(i);
    }

    return valuesArray;
}

/** Adapts the native JS Date for use
 * with cdk-based components that work with dates.
 */
@Injectable()
export class NativeDateAdapter extends DateAdapter<Date> {

    /** Whether to clamp the date between 1 and 9999 to avoid IE and Edge errors. */
    private readonly _clampDate: boolean;

    constructor(
        @Optional() @Inject(MC_DATE_LOCALE) mcDateLocale: string,
        platform: Platform
    ) {
        super();
        super.setLocale(mcDateLocale);

        this._clampDate = platform.TRIDENT || platform.EDGE;
    }

    getYear(date: Date): number {
        return date.getFullYear();
    }

    getMonth(date: Date): number {
        return date.getMonth();
    }

    getDate(date: Date): number {
        return date.getDate();
    }

    getDayOfWeek(date: Date): number {
        return date.getDay();
    }

    getMonthNames(style: 'long' | 'short' | 'narrow'): string[] {
        const dtf = new Intl.DateTimeFormat(this.locale, {month: style, timeZone: 'utc'});

        return range(MONTH_IN_YEAR, (i) =>
            this._stripDirectionalityCharacters(this._format(dtf, new Date(ANY_YEAR, i, 1))));
    }

    getDateNames(): string[] {
        const dtf = new Intl.DateTimeFormat(this.locale, {day: 'numeric', timeZone: 'utc'});

        return range(DAY_IN_MONTH, (i) => this._stripDirectionalityCharacters(
            this._format(dtf, new Date(ANY_YEAR, 0, i + 1))));
    }

    getDayOfWeekNames(style: 'long' | 'short' | 'narrow'): string[] {
        const dtf = new Intl.DateTimeFormat(this.locale, {weekday: style, timeZone: 'utc'});

        return range(DAY_IN_WEEK, (i) => this._stripDirectionalityCharacters(
            this._format(dtf, new Date(ANY_YEAR, 0, i + 1))));
    }

    getYearName(date: Date): string {
        const dtf = new Intl.DateTimeFormat(this.locale, {year: 'numeric', timeZone: 'utc'});

        return this._stripDirectionalityCharacters(this._format(dtf, date));
    }

    getFirstDayOfWeek(): number {
        // TODO
        // We can't tell using native JS Date what the first day of the week is, we default to Sunday.
        return 0;
    }

    getNumDaysInMonth(date: Date): number {
        return this.getDate(this._createDateWithOverflow(
            this.getYear(date), this.getMonth(date) + 1, 0));
    }

    clone(date: Date): Date {
        return this.createDate(this.getYear(date), this.getMonth(date), this.getDate(date));
    }

    createDate(year: number, month: number, date: number): Date {
        // Check for invalid month and date (except upper bound on date which we have to check after
        // creating the Date).
        if (month < 0 || month > (MONTH_IN_YEAR - 1)) {
            throw Error(`Invalid month index "${month}". Month index has to be between 0 and 11.`);
        }

        if (date < 1) {
            throw Error(`Invalid date "${date}". Date has to be greater than 0.`);
        }

        const result = this._createDateWithOverflow(year, month, date);
        // Check that the date wasn't above the upper bound for the month, causing the month to overflow
        if (result.getMonth() !== month) {
            throw Error(`Invalid date "${date}" for month with index "${month}".`);
        }

        return result;
    }

    today(): Date {
        return new Date();
    }

    parse(value: any): Date | null {
        // We have no way using the native JS Date to set the parse format or locale, so we ignore these
        // parameters.
        if (typeof value === 'number') {
            return new Date(value);
        }

        return value ? new Date(Date.parse(value)) : null;
    }

    // tslint:disable-next-line
    format(date: Date, displayFormat: Object): string {
        if (!this.isValid(date)) {
            throw Error('NativeDateAdapter: Cannot format invalid date.');
        }

        // On IE and Edge the i18n API will throw a hard error that can crash the entire app
        // if we attempt to format a date whose year is less than 1 or greater than 9999.
        if (this._clampDate && (date.getFullYear() < 1 || date.getFullYear() > MAX_YEAR)) {
            // tslint:disable-next-line
            date = this.clone(date);
            date.setFullYear(Math.max(1, Math.min(MAX_YEAR, date.getFullYear())));
        }

        // tslint:disable-next-line
        displayFormat = {...displayFormat, timeZone: 'utc'};

        const dtf = new Intl.DateTimeFormat(this.locale, displayFormat);

        return this._stripDirectionalityCharacters(this._format(dtf, date));
    }

    addCalendarYears(date: Date, years: number): Date {
        return this.addCalendarMonths(date, years * MONTH_IN_YEAR);
    }

    addCalendarMonths(date: Date, months: number): Date {
        let newDate = this._createDateWithOverflow(
            this.getYear(date), this.getMonth(date) + months, this.getDate(date));

        // It's possible to wind up in the wrong month if the original month has more days than the new
        // month. In this case we want to go to the last day of the desired month.
        // Note: the additional + 12 % 12 ensures we end up with a positive number, since JS % doesn't
        // guarantee this.
        if (this.getMonth(newDate) !==
            ((this.getMonth(date) + months) % MONTH_IN_YEAR + MONTH_IN_YEAR) % MONTH_IN_YEAR) {
            newDate = this._createDateWithOverflow(this.getYear(newDate), this.getMonth(newDate), 0);
        }

        return newDate;
    }

    addCalendarDays(date: Date, days: number): Date {
        return this._createDateWithOverflow(
            this.getYear(date), this.getMonth(date), this.getDate(date) + days);
    }

    toIso8601(date: Date): string {
        return [
            date.getUTCFullYear(),
            this._2digit(date.getUTCMonth() + 1),
            this._2digit(date.getUTCDate())
        ].join('-');
    }

    /**
     * Returns the given value if given a valid Date or null. Deserializes valid ISO 8601 strings
     * (https://www.ietf.org/rfc/rfc3339.txt) into valid Dates and empty string into null. Returns an
     * invalid date for all other values.
     */
    deserialize(value: any): Date | null {
        if (typeof value === 'string') {
            if (!value) {
                return null;
            }
            // The `Date` constructor accepts formats other than ISO 8601, so we need to make sure the
            // string is the right format first.
            if (ISO_8601_REGEX.test(value)) {
                const date = new Date(value);
                if (this.isValid(date)) {
                    return date;
                }
            }
        }

        return super.deserialize(value);
    }

    isDateInstance(obj: any) {
        return obj instanceof Date;
    }

    isValid(date: Date) {
        return !isNaN(date.getTime());
    }

    invalid(): Date {
        return new Date(NaN);
    }

    /** Creates a date but allows the month and date to overflow. */
    private _createDateWithOverflow(year: number, month: number, date: number) {
        const result = new Date(year, month, date);

        // We need to correct for the fact that JS native Date treats years in range [0, 99] as
        // abbreviations for 19xx.
        // tslint:disable-next-line
        if (year >= 0 && year < 100) {
            // tslint:disable-next-line
            result.setFullYear(this.getYear(result) - 1900);
        }

        return result;
    }

    /**
     * Strip out unicode LTR and RTL characters. Edge and IE insert these into formatted dates while
     * other browsers do not. We remove them to make output consistent and because they interfere with
     * date parsing.
     * @param str The string to strip direction characters from.
     * @returns The stripped string.
     */
    private _stripDirectionalityCharacters(str: string) {
        return str.replace(/[\u200e\u200f]/g, '');
    }

    /**
     * Pads a number to make it two digits.
     * @param n The number to pad.
     * @returns The padded number.
     */
    private _2digit(n: number) {
        // tslint:disable-next-line
        return ('00' + n).slice(-2);
    }

    /**
     * When converting Date object to string, javascript built-in functions may return wrong
     * results because it applies its internal DST rules. The DST rules around the world change
     * very frequently, and the current valid rule is not always valid in previous years though.
     * We work around this problem building a new Date object which has its internal UTC
     * representation with the local date and time.
     * @param dtf Intl.DateTimeFormat object, containg the desired string format. It must have
     *    timeZone set to 'utc' to work fine.
     * @param date Date from which we want to get the string representation according to dtf
     * @returns A Date object with its UTC representation based on the passed in date info
     */
    private _format(dtf: Intl.DateTimeFormat, date: Date) {
        const d = new Date(Date.UTC(
            date.getFullYear(), date.getMonth(), date.getDate(), date.getHours(),
            date.getMinutes(), date.getSeconds(), date.getMilliseconds()));

        return dtf.format(d);
    }

}
